+++ crates/weaver_resolver/src/registry.rs
@@ -814,6 +814,7 @@ mod tests {
                 registry_id,
                 &format!("{}/registry/*.yaml", test_dir),
             )
+            .into_result_failing_non_fatal()
             .expect("Failed to load semconv specs");
 
             let mut attr_catalog = AttributeCatalog::default();
@@ -874,6 +875,7 @@ mod tests {
         let mut sc_specs = SemConvRegistry::new("default");
         sc_specs
             .add_semconv_spec_from_string("<str>", registry_spec)
+            .into_result_failing_non_fatal()
             .expect("Failed to load semconv spec");
 
         let mut attr_catalog = AttributeCatalog::default();
@@ -1017,7 +1019,8 @@ groups:
         let mut semconv_registry = SemConvRegistry::try_from_path_pattern(
             registry_id,
             "data/registry-test-7-spans/registry/*.yaml",
-        )?;
+        )
+        .into_result_failing_non_fatal()?;
 
         // Resolve the semantic convention registry.
         let resolved_schema =