+++ crates/weaver_semconv/src/registry.rs
@@ -10,6 +10,7 @@ use crate::stats::Stats;
 use crate::Error;
 use std::collections::HashMap;
 use std::path::Path;
+use weaver_common::result::WResult;
 
 /// A semantic convention registry is a collection of semantic convention
 /// specifications indexed by group id.
@@ -64,20 +65,37 @@ impl SemConvRegistry {
     /// # Errors
     ///
     /// If the registry path pattern is invalid.
-    pub fn try_from_path_pattern(registry_id: &str, path_pattern: &str) -> Result<Self, Error> {
-        let mut registry = SemConvRegistry::new(registry_id);
-        for sc_entry in glob::glob(path_pattern).map_err(|e| Error::InvalidRegistryPathPattern {
-            path_pattern: path_pattern.to_owned(),
-            error: e.to_string(),
-        })? {
-            let path_buf = sc_entry.map_err(|e| Error::InvalidRegistryPathPattern {
-                path_pattern: path_pattern.to_owned(),
-                error: e.to_string(),
-            })?;
-            let semconv_spec = SemConvSpecWithProvenance::from_file(path_buf.as_path())?;
-            registry.add_semconv_spec(semconv_spec);
+    pub fn try_from_path_pattern(registry_id: &str, path_pattern: &str) -> WResult<Self, Error> {
+        fn create_registry_or_fatal(
+            registry_id: &str,
+            path_pattern: &str,
+            non_fatal_errors: &mut Vec<Error>,
+        ) -> Result<SemConvRegistry, Error> {
+            let mut registry = SemConvRegistry::new(registry_id);
+            for sc_entry in
+                glob::glob(path_pattern).map_err(|e| Error::InvalidRegistryPathPattern {
+                    path_pattern: path_pattern.to_owned(),
+                    error: e.to_string(),
+                })?
+            {
+                let path_buf = sc_entry.map_err(|e| Error::InvalidRegistryPathPattern {
+                    path_pattern: path_pattern.to_owned(),
+                    error: e.to_string(),
+                })?;
+                let (semconv_spec, nfes) = SemConvSpecWithProvenance::from_file(path_buf.as_path())
+                    .into_result_with_non_fatal()?;
+                registry.add_semconv_spec(semconv_spec);
+                non_fatal_errors.extend(nfes);
+            }
+            Ok(registry)
         }
-        Ok(registry)
+
+        let mut non_fatal_errors = vec![];
+
+        match create_registry_or_fatal(registry_id, path_pattern, &mut non_fatal_errors) {
+            Ok(registry) => WResult::with_non_fatal_errors(registry, non_fatal_errors),
+            Err(e) => WResult::FatalErr(e),
+        }
     }
 
     /// Creates a semantic convention registry from the given list of
@@ -121,9 +139,8 @@ impl SemConvRegistry {
     pub fn add_semconv_spec_from_file<P: AsRef<Path> + Clone>(
         &mut self,
         path: P,
-    ) -> Result<(), Error> {
-        self.add_semconv_spec(SemConvSpecWithProvenance::from_file(path.clone())?);
-        Ok(())
+    ) -> WResult<(), Error> {
+        SemConvSpecWithProvenance::from_file(path.clone()).map(|spec| self.add_semconv_spec(spec))
     }
 
     /// Load and add a semantic convention string to the semantic convention registry.
@@ -131,24 +148,22 @@ impl SemConvRegistry {
         &mut self,
         provenance: &str,
         spec: &str,
-    ) -> Result<(), Error> {
-        self.add_semconv_spec(SemConvSpecWithProvenance::from_string(provenance, spec)?);
-        Ok(())
+    ) -> WResult<(), Error> {
+        SemConvSpecWithProvenance::from_string(provenance, spec)
+            .map(|spec| self.add_semconv_spec(spec))
     }
 
     /// Loads and returns the semantic convention spec from a file.
     pub fn semconv_spec_from_file<P: AsRef<Path>>(
         semconv_path: P,
-    ) -> Result<(String, SemConvSpec), Error> {
+    ) -> WResult<(String, SemConvSpec), Error> {
         let provenance = semconv_path.as_ref().display().to_string();
-        let spec = SemConvSpec::from_file(semconv_path)?;
-        Ok((provenance, spec))
+        SemConvSpec::from_file(semconv_path).map(|spec| (provenance, spec))
     }
 
     /// Downloads and returns the semantic convention spec from an URL.
-    pub fn semconv_spec_from_url(sem_conv_url: &str) -> Result<(String, SemConvSpec), Error> {
-        let spec = SemConvSpec::from_url(sem_conv_url)?;
-        Ok((sem_conv_url.to_owned(), spec))
+    pub fn semconv_spec_from_url(sem_conv_url: &str) -> WResult<(String, SemConvSpec), Error> {
+        SemConvSpec::from_url(sem_conv_url).map(|spec| (sem_conv_url.to_owned(), spec))
     }
 
     /// Returns the number of semantic convention specs added in the semantic
@@ -204,12 +219,15 @@ mod tests {
     #[test]
     fn test_try_from_path_pattern() {
         // Test with a valid path pattern
-        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml").unwrap();
+        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml")
+            .into_result_failing_non_fatal()
+            .unwrap();
         assert_eq!(registry.id(), "test");
         assert_eq!(registry.semconv_spec_count(), 3);
 
         // Test with an invalid path pattern
-        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c***.yml");
+        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c***.yml")
+            .into_result_failing_non_fatal();
         assert!(registry.is_err());
         assert!(matches!(
             registry.unwrap_err(),
@@ -220,7 +238,8 @@ mod tests {
     #[test]
     fn test_semconv_spec_from_url() {
         let semconv_url = "https://raw.githubusercontent.com/open-telemetry/semantic-conventions/main/model/url.yaml";
-        let result = SemConvRegistry::semconv_spec_from_url(semconv_url);
+        let result =
+            SemConvRegistry::semconv_spec_from_url(semconv_url).into_result_failing_non_fatal();
         assert!(result.is_ok());
     }
 
@@ -303,19 +322,24 @@ mod tests {
 
     #[test]
     fn test_semconv_from_path_pattern() {
-        let mut registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml").unwrap();
+        let mut registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml")
+            .into_result_failing_non_fatal()
+            .unwrap();
         assert_eq!(registry.id(), "test");
         assert_eq!(registry.semconv_spec_count(), 3);
 
         registry
             .add_semconv_spec_from_file("data/database.yaml")
+            .into_result_failing_non_fatal()
             .unwrap();
         assert_eq!(registry.semconv_spec_count(), 4);
     }
 
     #[test]
     fn test_stats() {
-        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml").unwrap();
+        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml")
+            .into_result_failing_non_fatal()
+            .unwrap();
         let stats = registry.stats();
         assert_eq!(stats.file_count, 3);
         assert_eq!(stats.group_count, 3);
@@ -333,7 +357,9 @@ mod tests {
 
     #[test]
     fn test_unresolved_group_with_provenance_iter() {
-        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml").unwrap();
+        let registry = SemConvRegistry::try_from_path_pattern("test", "data/c*.yaml")
+            .into_result_failing_non_fatal()
+            .unwrap();
 
         let groups = registry
             .unresolved_group_with_provenance_iter()