+++ tests/resolution_process.rs
@@ -2,6 +2,8 @@
 
 //! Integration tests for the resolution process.
 
+use miette::Diagnostic;
+
 use weaver_cache::registry_path::RegistryPath;
 use weaver_cache::RegistryRepo;
 use weaver_common::TestLogger;
@@ -39,9 +41,12 @@ fn test_cli_interface() {
     let registry_repo = RegistryRepo::try_new("main", &registry_path).unwrap_or_else(|e| {
         panic!("Failed to create the registry repo, error: {e}");
     });
-    let semconv_specs = SchemaResolver::load_semconv_specs(&registry_repo).unwrap_or_else(|e| {
-        panic!("Failed to load the semantic convention specs, error: {e}");
-    });
+    let semconv_specs = SchemaResolver::load_semconv_specs(&registry_repo)
+        .ignore(|e| matches!(e.severity(), Some(miette::Severity::Warning)))
+        .into_result_failing_non_fatal()
+        .unwrap_or_else(|e| {
+            panic!("Failed to load the semantic convention specs, error: {e}");
+        });
     let semconv_specs = SemConvRegistry::from_semconv_specs(registry_id, semconv_specs);
 
     // Check if the logger has reported any warnings or errors.